Add day 16 part 2 solution
I had a really hard time with getting my solution to run fast enough
for this part 2. I abandoned my previous iterative idea for a
recursive memoizing approach but I'm not really sure that the
memoization was a useful addition. Runtime was about 5 minutes on my
machine and it consumed around 16-17GB of RAM. There must be a better
way.