day 16 part 2: more small optimizations
Now runs in about 28 seconds and takes about 1.5GB of RAM.

I think the main win here was from removing the initial vertex
from the neighbour lists of the complete graph. I accidentally left that
in when I was originally writing the thing, you only need to hold onto
the outgoing edges from it since it has no flow rate.